+++ lib/pipeline/application.ex
@@ -1,0 +1,11 @@
+defmodule Pipeline.Application do
+  use Application
+
+  def start(_type, _args) do
+    children = [
+      {Task.Supervisor, name: Pipeline.TaskSupervisor}
+    ]
+
+    Supervisor.start_link(children, strategy: :one_for_one)
+  end
+end